Move the database connection and operations to their own file
